Change in the filename usage, and few change added in StatusBar and menu items.
In this commit I've added procedure of filename read and write. Few methods checking if current project is saved and using name of active file to showing this on filename field in JFileChooser when client will want to again save this project.

Further, has been added few methods and action for menu items like:
 - showing or hiding status bar.
 - exchanging parameters MM0 and MM1.
 - reset project etc.